LayoutTests/platform/mac/media/media-can-play-wav-audio-expected.txt

Test HTMLMediaElement canPlayType() method with multiple .wav MIME types.

These tests may be expected to fail if the WebKit port does not support the format.

EXPECTED (audio.canPlayType('audio/wav') == 'maybe') OK
EXPECTED (audio.canPlayType('audio/x-wav') == 'maybe') OK
EXPECTED (audio.canPlayType('audio/wav; codecs=1') == 'probably'), OBSERVED 'maybe' FAIL
EXPECTED (audio.canPlayType('audio/x-wav; codecs=1') == 'probably'), OBSERVED 'maybe' FAIL
END OF TEST
